requirements.txt
tabulate
colorama
python-dotenv
requests 